test: add unit tests for new method [DHIS2-16290]
